feat: Create HuggingfaceSection component to encapsulate upload and download functionality for HuggingFace datasets in EditDatasetPage
